Container-WebApp/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
psycopg2-binary==2.9.10
email-validator==2.2.0
